Bug #41687 udn: Fix EOF handling
> notify.c:290:37: warning: comparison of constant 255 with expression of type 'char' is always true [-Wtautological-constant-out-of-range-compare]
>         } while (c != '\n' && c != -1 && c != 255 && ftell(notify->tf) != 1);
>                                          ~ ^  ~~~
> notify.c:292:19: warning: comparison of constant 255 with expression of type 'char' is always false [-Wtautological-constant-out-of-range-compare]
>         if (c == -1 || c == 255) {
>                        ~ ^  ~~~
> notify.c:372:9: warning: comparison of constant 255 with expression of type 'char' is always false [-Wtautological-constant-out-of-range-compare]
>                 if (c == 255) {
>                     ~ ^  ~~~

Reding <man:fgets(3)>:
> fgetc() ... returns ... an *unsigned char* cast to an *int*, or EOF ...

EOF is defined as (-1) in /usr/include/libio.h, which defaults to
"signed int" in C.

On x64 "char" is also a *signed*, which stores EOF:=(-1) as "-1".

But PowerPC64 defaults to *unsigned*: So EOF:=(-1) gets stored as
(unsigned char)-1 = 0xff = 255. When later that variable is comapred
again with EOF, the C compiler is comparing "signed int" with "unsigned
char" and needs to promote the later to the former. As +255 fits into
"int" nicely, the comparison then fails as (int)-1 != (int)255

<https://trofi.github.io/posts/203-signed-char-or-unsigned-char.html>

Using "char" is not enough to store the result of "fgetc()", so use
"int".