Bug #41687 udn: Improve file seeking
Just evaluate the return value of fseek() instead of calling ftell() for
each look iteration. Then the loop will never return EOF as we are never
reading the last character of the file as we start at least 3 bytes from
the end.

Just initialize the return variable with 0 and let sscanf() do the
parsing on success.